credit_scoring_app/requirements.txt
streamlit==1.28.0
pandas>=1.5.3
#numpy>=1.24.3
numpy==1.26.4  # 已更改为特定版本，通常会下载预编译的wheel
scikit-learn>=1.3.0
matplotlib>=3.7.2
seaborn>=0.12.2
shap==0.42.1
joblib>=1.3.2
pytest>=7.4.0
flask>=2.3.2
flask-cors>=4.0.0
sqlalchemy>=2.0.20
bcrypt>=4.0.1
python-dotenv>=1.0.0
pyjwt>=2.8.0
pydantic>=2.4.0
bleach>=6.0.0